requirements.txt
altgraph==0.17.4
bcrypt==4.2.0
cffi==1.17.1
cryptography==43.0.3
iniconfig==2.0.0
macholib==1.16.3
numpy==2.1.2
packaging==24.1
pandas==2.2.3
paramiko==3.5.0
pluggy==1.5.0
psycopg2==2.9.10
pycparser==2.22
pyinstaller==6.11.0
pyinstaller-hooks-contrib==2024.9
PyNaCl==1.5.0
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
setuptools==75.2.0
six==1.16.0
sshtunnel==0.4.0
tzdata==2024.2
wheel==0.44.0
